refactor particle filter node into separate class and node file
